+++ backend/Mloflo/products/admin.py
@@ -3,6 +3,4 @@ from .models import *
 
 admin.site.register(Product)
 #admin.site.register(ProductImage)
-admin.site.register(Order)
-admin.site.register(OrderItem)
-admin.site.register(ShippingAddress)
+
